Move Command up one namespace. Symfony Console is the defacto default for Silex/Laravel/Symfony/Composer/ezPublish/Drupal
