+++ src/Bernard/Command/ProduceCommand.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Bernard\Symfony\Command;
+namespace Bernard\Command;
 
 use Bernard\Producer;
 use Bernard\Message\DefaultMessage;
